+++ EduVPN-multiOS/Coordinators/AppCoordinator.swift
@@ -258,7 +258,6 @@ class AppCoordinator: RootViewCoordinator {
             .then { response -> Promise<CertificateModel> in response.mapResponse() }
             .map { model -> CertificateModel in
                 api.certificateModel = model
-                self.scheduleCertificateExpirationNotification(for: model, on: api)
                 return model
             }
     }
@@ -329,57 +328,6 @@ class AppCoordinator: RootViewCoordinator {
             #endif
         } else {
             showProfilesViewController(animated: animated)
-        }
-    }
-    
-    private func _scheduleCertificateExpirationNotification(for certificate: CertificateModel, on api: Api) {
-        guard let expirationDate = certificate.x509Certificate?.notAfter else { return }
-        guard let identifier = certificate.uniqueIdentifier else { return }
-        
-        let notificationTitle = NSLocalizedString("VPN certificate is expiring", comment: "")
-        
-        var notificationBody: String?
-        if let certificateTitle = api.instance?.displayNames?.localizedValue {
-            notificationBody = String.localizedStringWithFormat("Once expired the certificate for instance %@ needs to be refreshed.", certificateTitle)
-        }
-        
-        let notification = NotificationsService.makeNotification(title: notificationTitle, body: notificationBody)
-        
-        #if DEBUG
-        
-        guard let expirationWarningDate = NSCalendar.current.date(byAdding: .second, value: 10, to: Date()) else { return }
-        let expirationWarningDateComponents = NSCalendar.current.dateComponents(in: NSTimeZone.default, from: expirationWarningDate)
-        
-        #else
-        
-        guard let expirationWarningDate = (expirationDate.timeIntervalSinceNow < 86400 * 7) ? (NSCalendar.current.date(byAdding: .day, value: -7, to: expirationDate)) : (NSCalendar.current.date(byAdding: .minute, value: 10, to: Date())) else { return }
-        
-        var expirationWarningDateComponents = NSCalendar.current.dateComponents(in: NSTimeZone.default, from: expirationWarningDate)
-        
-        // Configure the trigger for 10am.
-        expirationWarningDateComponents.hour = 10
-        expirationWarningDateComponents.minute = 0
-        expirationWarningDateComponents.second = 0
-        
-        #endif
-
-        os_log("Scheduling a cert expiration reminder for %{public}@ on %{public}@.", log: Log.general, type: .info, certificate.uniqueIdentifier ?? "", expirationDate.description)
-        NotificationsService.sendNotification(notification, withIdentifier: identifier, at: expirationWarningDateComponents) { error in
-            if let error = error {
-                os_log("Error occured when scheduling a cert expiration reminder %{public}@",
-                       log: Log.general,
-                       type: .info, error.localizedDescription)
-            }
-        }
-    }
-    
-    fileprivate func scheduleCertificateExpirationNotification(for certificate: CertificateModel, on api: Api) {
-        NotificationsService.permissionGranted {
-            if $0 {
-                self._scheduleCertificateExpirationNotification(for: certificate, on: api)
-            } else {
-                os_log("Not Authorised", log: Log.general, type: .info)
-            }
         }
     }